aded in policy scoping for most functions. Create a profile needs correcting as currently the method for the user only having one profile doesnt work due to access to all profiles being used to validate the user in the front end

+++ backend/app/controllers/api/v1/profiles_controller.rb
@@ -1,19 +1,28 @@
-class Api::V1::ProfilesController < ApplicationController
+class Api::V1::ProfilesController < ApiController
   before_action :set_profile, only: [:show, :update, :destroy]
   def index
-    @profiles = Profile.all
+    @profiles = policy_scope(Profile)
+
+    # @profiles = Profile.where(user_id: current_user.id)
+    # @profiles = current_user.profiles
     render json: @profiles, status: :ok
-      # @profiles = current_user.profiles
   end
 
   def show
     render json: @profile, status: :ok
   end
 
+  def new
+    @profile = Profile.new
+    authorize @profile
+  end
+
   def create
     @profile = Profile.new(profile_params)
     # @profile = current_user.profile.new(company_params)
+    puts "current user #{current_user}"
     @profile.user_id = current_user.id
+    authorize @profile
     if @profile.save
       render json: @profile, status: :ok
     else
@@ -22,7 +31,14 @@ class Api::V1::ProfilesController < ApplicationController
     end
   end
 
+  def edit
+    authorize @profile
+
+  end
+
   def update
+    authorize @profile
+
     if @profile.update(profile_params)
       render json: @profile, status: :ok
     else
@@ -32,6 +48,8 @@ class Api::V1::ProfilesController < ApplicationController
   end
 
   def destroy
+    authorize @profile
+
     if @profile.destroy
       render json: { data: 'Profile deleted successfully', status: 'success'},
       status: :ok
